Initial code to get workspaces rendering in the navigator
This implementation is still buggy.  The two bugs encountered thus far
are that 1) workspaces only render properly on the first attempt to
open them and 2) occasionally, workspaces render in a disabled state.
The second bug is rare and I've yet to figure out what causes it.

In addition, existing modules need to be modified before they will
function with the new system.  Namely, you _must_ add workspaces to the
list of panels.  For compatability reasons, they _must_ be after the
existing panels.  Further, in order to not display in the sidebar, they
must also have the key-value pair `workspace: true`.

Currently, only the real_estate module functions with the new workspace system.